consistently rename feature-flagged modules
