+++ diman/src/rand.rs
